+++ packages/tailwindcss/src/index.test.ts
@@ -115,6 +115,34 @@ describe('compiling CSS', () => {
       ),
     ).toMatchSnapshot()
   })
+
+  test('unescapes underscores to spaces inside arbitrary values except for `url()` and first argument of `var()`', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme {
+            --spacing-1_5: 1.5rem;
+            --spacing-2_5: 2.5rem;
+          }
+          @tailwind utilities;
+        `,
+        ['bg-[no-repeat_url(./my_file.jpg)', 'ml-[var(--spacing-1_5,_var(--spacing-2_5,_1rem))]'],
+      ),
+    ).toMatchInlineSnapshot(`
+      ":root {
+        --spacing-1_5: 1.5rem;
+        --spacing-2_5: 2.5rem;
+      }
+
+      .ml-\\[var\\(--spacing-1_5\\,_var\\(--spacing-2_5\\,_1rem\\)\\)\\] {
+        margin-left: var(--spacing-1_5, var(--spacing-2_5, 1rem));
+      }
+
+      .bg-\\[no-repeat_url\\(\\.\\/my_file\\.jpg\\) {
+        background-color: no-repeat url("./")my file. jpg;
+      }"
+    `)
+  })
 })
 
 describe('arbitrary properties', () => {
@@ -670,6 +698,100 @@ describe('sorting', () => {
         syntax: "<number>";
         inherits: false;
         initial-value: 0;
+      }"
+    `)
+  })
+
+  it('should sort individual logical properties later than left/right pairs', async () => {
+    expect(
+      await compileCss(
+        css`
+          @theme {
+            --spacing-1: 1px;
+            --spacing-2: 2px;
+            --spacing-3: 3px;
+          }
+          @tailwind utilities;
+        `,
+        [
+          // scroll-margin
+          'scroll-ms-1',
+          'scroll-me-2',
+          'scroll-mx-3',
+
+          // scroll-padding
+          'scroll-ps-1',
+          'scroll-pe-2',
+          'scroll-px-3',
+
+          // margin
+          'ms-1',
+          'me-2',
+          'mx-3',
+
+          // padding
+          'ps-1',
+          'pe-2',
+          'px-3',
+        ].sort(() => Math.random() - 0.5),
+      ),
+    ).toMatchInlineSnapshot(`
+      ":root {
+        --spacing-1: 1px;
+        --spacing-2: 2px;
+        --spacing-3: 3px;
+      }
+
+      .mx-3 {
+        margin-left: var(--spacing-3, 3px);
+        margin-right: var(--spacing-3, 3px);
+      }
+
+      .ms-1 {
+        margin-inline-start: var(--spacing-1, 1px);
+      }
+
+      .me-2 {
+        margin-inline-end: var(--spacing-2, 2px);
+      }
+
+      .scroll-mx-3 {
+        scroll-margin-left: var(--spacing-3, 3px);
+        scroll-margin-right: var(--spacing-3, 3px);
+      }
+
+      .scroll-ms-1 {
+        scroll-margin-inline-start: var(--spacing-1, 1px);
+      }
+
+      .scroll-me-2 {
+        scroll-margin-inline-end: var(--spacing-2, 2px);
+      }
+
+      .scroll-px-3 {
+        scroll-padding-left: var(--spacing-3, 3px);
+        scroll-padding-right: var(--spacing-3, 3px);
+      }
+
+      .scroll-ps-1 {
+        scroll-padding-inline-start: var(--spacing-1, 1px);
+      }
+
+      .scroll-pe-2 {
+        scroll-padding-inline-end: var(--spacing-2, 2px);
+      }
+
+      .px-3 {
+        padding-left: var(--spacing-3, 3px);
+        padding-right: var(--spacing-3, 3px);
+      }
+
+      .ps-1 {
+        padding-inline-start: var(--spacing-1, 1px);
+      }
+
+      .pe-2 {
+        padding-inline-end: var(--spacing-2, 2px);
       }"
     `)
   })